hospital/pyproject.toml
[tool.poetry]
name = "hospitals"
version = "0.1.0"
description = ""
authors = ["ligets <legen2a208@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.0"
uvicorn = "^0.31.0"
alembic = "^1.13.3"
asyncpg = "^0.29.0"
pydantic = "^2.9.2"
pydantic-settings = "^2.5.2"
python-dotenv = "^1.0.1"
datetime = "^5.5"
pyjwt = "^2.9.0"
aio-pika = "^9.4.3"
fastapi-cache2 = {extras = ["redis"], version = "^0.2.2"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
